feat: split controllers & added docs
