+++ apps/github-node/src/controllers/trigger-controller.js
@@ -1,7 +1,13 @@
 const subscriptionModelService = require('../services/model-services/subscription-model-service');
 const connectionModelService = require('../services/model-services/connection-model-service');
 const githubService = require('../services/github-service');
+const mondayTriggersService = require('../services/monday-triggers-service');
 
+/**
+ * Creates a Subscription and Github webhook. 
+ * Called when a user adds the integration to their board.
+ * Docs: https://developer.monday.com/apps/docs/custom-trigger#subscribing-to-your-trigger
+ */
 async function subscribe(req, res) {
   const { userId } = req.session;
   const { payload } = req.body;
@@ -29,6 +35,11 @@ async function subscribe(req, res) {
   }
 }
 
+/**
+ * Removes the Subscription and webhook associated with a specific integration. 
+ * Called when a user deletes or turns off the integration on their board.
+ * Docs: https://developer.monday.com/apps/docs/custom-trigger#unsubscribing-from-your-trigger
+ */
 async function unsubscribe(req, res) {
   const { userId } = req.session;
   const { webhookId: subscriptionId } = req.body.payload;
@@ -46,42 +57,37 @@ async function unsubscribe(req, res) {
   }
 }
 
-async function executeAction(req, res) {
-  const { userId } = req.session;
-  const { inputFields } = req.body.payload;
+/**
+ * Trigger the integration recipe by receive events from Github and then calling the corresponding monday webhook URL. 
+ * Docs: https://developer.monday.com/apps/docs/custom-trigger#calling-your-action
+ */
+async function integrationEventsHandler(req, res) {
+  const { subscriptionId } = req.params;
+  const body = req.body;
+
+  if (body.hook_id) {
+    // Initial URL verification step from Github
+    return res.status(200).send({ success: true });
+  }
 
   try {
-    const { token } = await connectionModelService.getConnectionByUserId(userId);
+    const { action, issue } = body;
+    if (action != 'opened') {
+      // Only invoke trigger when an issue is opened
+      return res.status(200).send();
+    }
 
-    const { repository, issue } = inputFields;
-    const owner = repository.value.owner;
-    const repo = repository.value.name;
+    // Get the Subscription object and the corresponding monday webhook URL
+    const { mondayWebhookUrl } = await subscriptionModelService.getSubscription(subscriptionId);
 
-    await githubService.createIssue(token, owner, repo, issue);
+    // Converts the incoming issue to mappable fields in monday
+    // Docs: https://developer.monday.com/apps/docs/dynamic-mapping#part-3-using-inbound-mapping-your-app-to-mondaycom
+    const convertedIssue = githubService.convertIssueToPrimitives(issue);
+
+    // Call the monday webhook URL
+    await mondayTriggersService.triggerMondayIntegration(mondayWebhookUrl, { issue: convertedIssue });
 
     return res.status(200).send();
-  } catch (err) {
-    console.error(err);
-    return res.status(500).send({ message: 'internal server error' });
-  }
-}
-
-async function getRemoteListOptions(req, res) {
-  const { userId } = req.session;
-  try {
-    const { token } = await connectionModelService.getConnectionByUserId(userId);
-    const options = await githubService.getRepositories(token);
-    return res.status(200).send(options);
-  } catch (err) {
-    console.error(err);
-    return res.status(500).send({ message: 'internal server error' });
-  }
-}
-
-async function getFieldDefs(req, res) {
-  try {
-    const fieldDefs = githubService.getIssueFieldDefs();
-    return res.status(200).send(fieldDefs);
   } catch (err) {
     console.error(err);
     return res.status(500).send({ message: 'internal server error' });
@@ -91,7 +97,5 @@ async function getFieldDefs(req, res) {
 module.exports = {
   subscribe,
   unsubscribe,
-  executeAction,
-  getRemoteListOptions,
-  getFieldDefs,
-};
+  integrationEventsHandler,
+}
